+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### What's new
 
 * New feature: Edit pre-written comments after their publication with the advanced publishing settings menu. Comments links are now stored in the `tg_comments` property. Note: `telegram_links` property was renamed to `tg_posts`. If you used feature of saving posts links to properties, be sure to change the name of already existing property with Obsidian core plugin [Properties view](https://obsidian.md/help/plugins/properties).
-* New feature: View changelog in the settings, in the user guide, or via palette command.
+* New feature: View changelog in the settings, in the user guide, or via the palette command.
 
 ### UI/UX enhancements
 
@@ -16,6 +16,7 @@
 ### Other
 
 * GramJS localStorage API schema cache was disabled to avoid triggering Obsidian linter.
+* Fixed UI bugs in the advanced settings.
 * Various markdown parsing fixes.